+++ src/options.h
@@ -65,6 +65,11 @@ struct program_options {
 #define OPT_ADDSLASH                1
 #define DFLT_OPT_ADDSLASH           OPT_NOADDSLASH
     unsigned char add_slash;
+/* add parent directories (option -P) */
+#define OPT_NOADDPARENTS            0
+#define OPT_ADDPARENTS              1
+#define DFLT_OPT_ADDPARENTS         OPT_NOADDPARENTS
+    unsigned char add_parents;
 /* verbose output (option -v) */
 #define OPT_NOVERBOSE               0
 #define OPT_VERBOSE                 1